backend/requirements.txt
# ШІ-Агент Контактного Центру - Backend Dependencies

# Web Framework
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9

# WebSocket support
websockets==12.0

# ASR - Silero Speech-to-Text
silero==0.4.0
torch>=2.0.0
torchaudio>=2.0.0

# TTS - Fish Speech
# fish-speech (install from git: pip install git+https://github.com/fishaudio/fish-speech.git)

# NLU / Classifier
transformers>=4.35.0
sentence-transformers>=2.2.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
numpy>=1.24.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Oracle APEX integration
cx_Oracle==8.3.0
oracledb==2.0.1

# Utilities
python-dotenv==1.0.0
httpx==0.27.0
pydantic==2.6.0
